directory restructuring
to address potential interoperability conflicts with different host programs, as well as an irritating tendency for people to muck about in mfvitools code to no purpose
no longer assuming a BC EX directory structure with johnnydmad files in root. now johnnydmad is designed to live in a subdirectory with its own consistent structure independent of host.