experiment: motif
trying out a new mode that tries to restrict the number of games/sources for music, hoping for a result somewhere in the general ballpark of "use every chrono trigger song available, then fill the rest with FF5" or something along those lines. predictably, like the similar "cloneparty" BC tag, it doesn't actually work very well and probably will remain an experiment.

a key problem is that since there are pretty few games with large number of tracks, the results are somewhere between "can't tell anything's different" and "this is the same as the last two times."