replace dc2_garden.mml
the story here is that Cecil188's version of this track was based on a score shown in a youtube video, and the poster of that youtube video admitted in a comment that they hadn't created the MIDI file it was displaying. thus, we can't determine who to rightfully credit for the score. So, it's replaced with a new version from scratch with no ambiguous credits.